+++ src/config-type.ts
@@ -1,5 +1,6 @@
 import {NextHandleFunction} from 'connect';
-import * as winston from 'winston';
+import {Express} from 'express';
+import {Logger} from 'winston';
 import {ResultHandler} from './result-handler';
 
 export const loggerLevels = {
@@ -13,18 +14,21 @@ export interface LoggerConfig {
   color: boolean;
 }
 
-export interface ConfigType {
-  server: {
+export type ConfigType = ({
+  server: { // server configuration
     // port or socket
     listen: number | string;
-    // enable cross-origin resource sharing
-    // https://developer.mozilla.org/en-US/docs/Web/HTTP/CORS
-    cors: boolean;
     // custom JSON parser, default: express.json()
-    jsonParser?: NextHandleFunction,
-    // custom handler for errors and output, default: defaultResultHandler()
-    resultHandler?: ResultHandler;
+    jsonParser?: NextHandleFunction;
   },
+} | { // or your custom express app
+  app: Express
+}) & {
+  // enable cross-origin resource sharing
+  // https://developer.mozilla.org/en-US/docs/Web/HTTP/CORS
+  cors: boolean;
+  // custom handler for errors and output, default: defaultResultHandler()
+  resultHandler?: ResultHandler;
   // logger configuration or your custom winston logger
-  logger: LoggerConfig | winston.Logger;
+  logger: LoggerConfig | Logger;
 }